Final commit

+++ my_aes.c++
@@ -3,7 +3,7 @@
 
 // MyAES Constructors
 MyAES::MyAES() {}
-MyAES::MyAES(const int& _key_size,
+MyAES::MyAES(const uint16_t& _key_size,
              const std::string& _key_file,
              const std::string& _input_file,
              const std::string& _output_file,
@@ -32,7 +32,7 @@ MyAES::MyAES(const int& _key_size,
 
 // Private/Helper Methods
 void MyAES::CheckPadding() {
-  int row, col;
+  int8_t row, col;
   row = col = 3;
   pad_size = 0;
   while (col >= 0 && data[row][col] == 0) {
@@ -45,7 +45,7 @@ void MyAES::CheckPadding() {
   }
 }
 void MyAES::FillInitVector() {
-  int i;
+  uint8_t i;
   char c;
   for (i = 0; i < 16 && key_file.get(c); ++i) {
     init_vector[i] = c;
@@ -55,7 +55,7 @@ void MyAES::FillInitVector() {
 }
 void MyAES::FillData() {
   char c;
-  int row, col;
+  uint8_t row, col;
   row = col = data_size = 0;
   // Read data from input file (one char/byte at a time)
   while (data_size < 16 && in_file.get(c)) {
@@ -82,45 +82,45 @@ void MyAES::FillData() {
   }
 }
 void MyAES::SubBytes() {
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
+  for (uint8_t i = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j) {
       data[i][j] = s[data[i][j]];
     }
   }
 }
 void MyAES::InvSubBytes() {
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
+  for (uint8_t i = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j) {
       data[i][j] = inv_s[data[i][j]];
     }
   }
 }
 void MyAES::ShiftLeft(byte* row) {
   byte temp = row[0];
-  for (int i = 0; i < 3; ++i)
+  for (uint8_t i = 0; i < 3; ++i)
     row[i] = row[i + 1];
   row[3] = temp;
 }
 void MyAES::ShiftRows() {
-  for (int i = 1; i <= 3; ++i) {
-    for (int j = 1; j <= i; ++j) {
+  for (uint8_t i = 1; i <= 3; ++i) {
+    for (uint8_t j = 1; j <= i; ++j) {
       ShiftLeft(data[i]);
     }
   }
 }
 void MyAES::InvShiftRows() {
-  for (int i = 1; i <= 3; ++i) {
-    for (int j = 3; j >= i; --j) {
+  for (uint8_t i = 1; i <= 3; ++i) {
+    for (uint8_t j = 3; j >= i; --j) {
       ShiftLeft(data[i]);
     }
   }
 }
 void MyAES::MixColumns() {
   byte temp[4][4] = {{0,0,0,0},{0,0,0,0},{0,0,0,0},{0,0,0,0}};
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
-      for (int k = 0; k < 4; ++k) {
-        int a = galois_matrix[j][k];
+  for (uint8_t i = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j) {
+      for (uint8_t k = 0; k < 4; ++k) {
+        uint8_t a = galois_matrix[j][k];
         byte b = data[k][i];
         if (a == 1) temp[j][i] ^= b;
         if (a == 2) temp[j][i] ^= two[b];
@@ -128,18 +128,18 @@ void MyAES::MixColumns() {
       }
     }
   }
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
+  for (uint8_t i = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j) {
       data[i][j] = temp[i][j];
     }
   }
 }
 void MyAES::InvMixColumns() {
   byte temp[4][4] = {{0,0,0,0},{0,0,0,0},{0,0,0,0},{0,0,0,0}};
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
-      for (int k = 0; k < 4; ++k) {
-        int a = inv_galois_matrix[j][k];
+  for (uint8_t i = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j) {
+      for (uint8_t k = 0; k < 4; ++k) {
+        uint8_t a = inv_galois_matrix[j][k];
         byte b = data[k][i];
         if (a == 9) temp[j][i] ^= nine[b];
         if (a == 11) temp[j][i] ^= eleven[b];
@@ -148,44 +148,44 @@ void MyAES::InvMixColumns() {
       }
     }
   }
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
+  for (uint8_t i = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j) {
       data[i][j] = temp[i][j];
     }
   }
 }
-void MyAES::GenerateKeyHelper(byte in[], int i) {
+void MyAES::GenerateKeyHelper(byte in[], uint8_t i) {
   ShiftLeft(in);
-  for (int a = 0; a < 4; ++a) {
+  for (uint8_t a = 0; a < 4; ++a) {
     in[a] = s[in[a]];
   }
   in[0] ^= rcon[i];
 }
-void MyAES::AddRoundKey(const int& round) {
-  for (int i = 0, k = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j, ++k) {
+void MyAES::AddRoundKey(const uint8_t& round) {
+  for (uint8_t i = 0, k = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j, ++k) {
       data[j][i] = data[j][i] ^ expanded_keys[round*16 + k];
     }
   }
 }
 void MyAES::StoreData() {
   CheckPadding();
-  for (int i = 0, k = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j, ++k) {
+  for (uint8_t i = 0, k = 0; i < 4; ++i) {
+    for (uint8_t j = 0; j < 4; ++j, ++k) {
       if (k < data_size - pad_size) out_file << data[j][i];
     }
   }
 }
 void MyAES::CopyData(std::vector<byte>& v) {
-  for (int i = 0, k = 0; i < 4; ++i) { 
-    for (int j = 0; j < 4; ++j, ++k) {
+  for (uint8_t i = 0, k = 0; i < 4; ++i) { 
+    for (uint8_t j = 0; j < 4; ++j, ++k) {
       v[k] = data[j][i];
     }
   }
 }
 void MyAES::XorDataCBC() {
-  for (int i = 0, k = 0; i < 4; ++i) { 
-    for (int j = 0; j < 4; ++j, ++k) {
+  for (uint8_t i = 0, k = 0; i < 4; ++i) { 
+    for (uint8_t j = 0; j < 4; ++j, ++k) {
       data[j][i] ^= cbc_buffer[k];
     }
   }
@@ -204,12 +204,12 @@ void MyAES::GenerateKeys() {
   // file. The first 16 bytes of the keyfile must represent the init. vector).
   if (cbc_mode) FillInitVector();
 
-  int n = (key_size == 128 ?  16 :  32);
+  uint8_t n = (key_size == 128 ?  16 :  32);
   byte temp[4];
   char x;
   // Get first 'n' bytes from the original key
   {
-    int i;
+    uint8_t i;
     for (i = 0; i < n && key_file.get(x); ++i) {
       expanded_keys[i] = x;
     }
@@ -219,11 +219,11 @@ void MyAES::GenerateKeys() {
     }
   }
   // Fill the remaining bytes using the specified iterative process
-  for (int processed_bytes = n, it = 1; processed_bytes < expanded_keys.size();
-       processed_bytes += 4) {
+  for (uint8_t processed_bytes = n, it = 1; 
+       processed_bytes < expanded_keys.size(); processed_bytes += 4) {
     // Assign the value of the previous 4 bytes to 'temp'. We are
     // going to build the next 4 bytes of our expanded key in 'temp'.
-    for (int i = 0; i < 4; ++i) {
+    for (uint8_t i = 0; i < 4; ++i) {
       temp[i] = expanded_keys[processed_bytes + i - 4];
     }
     // Every 'n' bytes (size each key) we want to re-
@@ -234,13 +234,13 @@ void MyAES::GenerateKeys() {
     }
     // We need to use the fixed s-table for 256-bit keys to make a substitution
     if (key_size == 256 && processed_bytes % n == (n >> 1)) {
-      for (int i = 0; i < 4; ++i) {
+      for (uint8_t i = 0; i < 4; ++i) {
         temp[i] = s[temp[i]];
       }
     }
     // Last Setp: XOR with expanded_keys that are 'n' bytes behind and
     // finally store the temp key that we've built so far in expanded_keys.
-    for (int i = 0; i < 4; ++i) {
+    for (uint8_t i = 0; i < 4; ++i) {
       temp[i] ^= expanded_keys[processed_bytes + i - n];
       expanded_keys[processed_bytes + i] = temp[i];
     }
@@ -251,19 +251,18 @@ void MyAES::Encrypt() {
   if (cbc_mode) std::copy(init_vector.begin(), init_vector.end(), 
                           cbc_buffer.begin());
   // Get the appropriate number of rounds needed depending on key size
-  int num_rounds = (key_size == 128) ? 10 : 14;
+  uint8_t num_rounds = (key_size == 128) ? 10 : 14;
   while (1) {
     // Get next 16 bytes of data from input file
     FillData();
     // Stop if no data was extracted (i.e. EOF)
     if (data_size == 0) break;
-    
     // XOR the 16 bytes of data with whatever is in 'cbc_buffer'
     if (cbc_mode) XorDataCBC();
-
+    
     // Do iterative process to encrypt data:
     AddRoundKey(0);
-    int round;
+    uint8_t round;
     for (round = 1; round < num_rounds; ++round) {
       SubBytes();
       ShiftRows();
@@ -273,6 +272,7 @@ void MyAES::Encrypt() {
     SubBytes();
     ShiftRows();
     AddRoundKey(round);
+
     // Store encrypted data to output file
     StoreData();
     // Copy the 16 bytes of ciphertext to 'cbc_buffer'
@@ -280,19 +280,18 @@ void MyAES::Encrypt() {
   }
 }
 void MyAES::Decrypt() {
-  // 'temp' is only used for CBC mode
+  // 'temp' is only used for CBC mode; we initialize it to the init. vector
   std::vector<byte> temp(init_vector);
   while (1) {
     // Copy whatever is in 'temp' into the cbc_buffer.
     if (cbc_mode) std::copy(temp.begin(), temp.end(), 
                             cbc_buffer.begin());
     // Get the appropriate number of rounds needed depending on key size
-    int round = (key_size == 128) ? 10 : 14;
+    uint8_t round = (key_size == 128) ? 10 : 14;
     // Get next 16 bytes of data from input file
     FillData();
     // Stop if no data was extracted (i.e. EOF)
     if (data_size == 0) break;
-    
     // Copy the 16 bytes of data into 'temp'
     if (cbc_mode) CopyData(temp);
